refactor(test): split SCIM handler tests into a per-topic directory
handlers/scim/tests.rs had grown past four thousand lines. Split it
into tests/{protocol,users,filters,validation,groups,meta}.rs with the
module root holding the shared import globs and a doc listing each
file's scope, mirroring handlers/oidc/tests/ and db/tests/. Also drop
two internal phase labels from section banners on the way through.

Test bodies are unchanged; the suite still runs 104 tests.
